+++ app/Evil Cave/Items/LWFItem.m
@@ -132,7 +132,7 @@
         labelColor = [UIColor colorWithRed:1.0 green:215.0/255.0 blue:0.0 alpha:1.0];
 
         [label setFontColor:labelColor];
-        label.text = [NSString stringWithFormat:@"%d", self.quantity];
+        label.text = [NSString stringWithFormat:@"%ld", (long)self.quantity];
         
         return label;
